+++ src/test/java/roomescape/JpaTest.java
@@ -29,21 +29,5 @@ public class JpaTest {
         assertThat(persistTime.getTime()).isEqualTo(time.getTime());
     }
 
-    @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-    @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-    public class MissionStepTest {
-        @Test
-        void 오단계() {
-            String adminToken = createToken("admin@email.com", "password");
 
-            List<MyReservationResponse> reservations = RestAssured.given().log().all()
-                    .cookie("token", adminToken)
-                    .get("/reservations-mine")
-                    .then().log().all()
-                    .statusCode(200)
-                    .extract().jsonPath().getList(".", MyReservationResponse.class);
-
-            assertThat(reservations).hasSize(3);
-        }
-    }
 }
